backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["CristiPV, ithai5"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.82.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
pytest = "^7.1.3"
requests = "^2.28.1"
beanie = "^1.11.9"
python-dotenv = "^0.21.0"
pytest-asyncio = "^0.19.0"
trio = "^0.22.0"
httpx = "^0.23.0"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
finnhub-python = "^2.4.15"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
